[bug] Fix CSV Injection (aka Formula Injection).
This is a CSV design flaw, not exactly a plugin bug. In short, if a field has a formula on it (ex. "=1+2"), Excel/OpenCalc will execute it and show the result (ex. "3"). In more extreme cases, an attacker can make Excel/Calc call an URL once the sheet is opened, which can cause them to know the victim’s IP. Most spreadsheet clients (Excel/Calc/Google Docs) have protections against this, asking for permission before executing any formula that can compromise the user. To totally mitigate this, we have implemented the OWASP recommendation for this issue which is to ensure that no cells begin with any of the following characters: Equals to (“=”), Plus (“+”), Minus (“-“), At (“@”).